LayoutTests/imported/w3c/web-platform-tests/css/css-ruby/br-clear-all-002-expected.txt

The yellow box should encompass its contents and also the cyan float, due to clearance. The yellow box has bottom padding, and the blue ruby annotation box is allowed to overflow into the padding area. In this case the blue box is so tall it will use the entire padding area, and also stretch the yellow box somewhat.



FAIL #container 1 assert_equals:
<div id="container" data-expected-height="170">
  <ruby>
    <div></div>
    <rt>
      <div></div>
    </rt>
  </ruby>
  <br clear="all">
</div>
height expected 170 but got 221
